Move PeriodicFixedDelayTask to utils
